+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "2"
 
 members = [
     "common",
+    "common-token",
     "deployer",
     "pool",
     "pool-interface",
