+++ common-token/src/lib.rs
@@ -1,0 +1,18 @@
+#![deny(warnings)]
+#![no_std]
+
+use soroban_sdk::{Address, Env};
+
+pub mod storage;
+
+pub fn verify_caller_is_pool(e: &Env) -> Address {
+    let pool = crate::storage::read_pool(e);
+    pool.require_auth();
+    pool
+}
+
+pub fn check_nonnegative_amount(amount: i128) {
+    if amount < 0 {
+        panic!("negative amount is not allowed: {}", amount)
+    }
+}
